Crate rustc_builtin_macros

This crate contains implementations of built-in macros and other code generating facilities injecting code into the crate before it is lowered to HIR.

Modules

alloc_error_handler ğŸ”’

asm

assert ğŸ”’

cfg ğŸ”’

The compiler code necessary to support the cfg! extension, which expands to a literal true or false based on whether the given cfg matches the current compilation environment.

cfg_accessible ğŸ”’

Implementation of the #[cfg_accessible(path)] attribute macro.

cfg_eval ğŸ”’

cmdline_attrs

Attributes injected into the crate root from command line using -Z crate-attr.

compile_error ğŸ”’

concat ğŸ”’

concat_bytes ğŸ”’

concat_idents ğŸ”’

derive ğŸ”’

deriving ğŸ”’

The compiler code necessary to implement the #[derive] extensions.

edition_panic ğŸ”’

env ğŸ”’

format ğŸ”’

format_foreign ğŸ”’

global_allocator ğŸ”’

log_syntax ğŸ”’

proc_macro_harness

source_util ğŸ”’

standard_library_imports

test ğŸ”’

test_harness

trace_macros ğŸ”’

type_ascribe ğŸ”’

util ğŸ”’

Functions

register_builtin_macros