rustc_builtin_macros::autodiff

Module llvm_enzyme

Functions

assure_mut_ref ğŸ”’

call_autodiff ğŸ”’

expand_forward ğŸ”’

expand_reverse ğŸ”’

expand_with_mode ğŸ”’

We expand the autodiff macro to generate a new placeholder function which passes type-checking and can be called by users. The exact signature of the generated function depends on the configuration provided by the user, but here is an example:

extract_item_info ğŸ”’

first_ident ğŸ”’

from_ast ğŸ”’

gen_enzyme_decl ğŸ”’

gen_turbofish_expr ğŸ”’

has_ret ğŸ”’

meta_item_inner_to_ts ğŸ”’

name ğŸ”’

outer_normal_attr ğŸ”’

width ğŸ”’