Enum rustc_builtin_macros::format_foreign::printf::Substitution

pub enum Substitution<'a> {
    Format(Format<'a>),
    Escape((usize, usize)),
}

Represents a single printf-style substitution.

Variants

Format(Format<'a>)

A formatted output substitution with its internal byte offset.

Escape((usize, usize))

A literal %% escape, with its start and end indices.

Implementations

impl<'a> Substitution<'a>

pub fn as_str(&self) -> &str

pub fn position(&self) -> Option<InnerSpan>

pub fn set_position(&mut self, start: usize, end: usize)

pub fn translate(&self) -> Result<String, Option<String>>

Translate this substitution into an equivalent Rust formatting directive.

This ignores cases where the substitution does not have an exact equivalent, or where the substitution would be unnecessary.