rustc_builtin_macros::test_harness
Function inject

pub fn inject(
    krate: &mut Crate,
    sess: &Session,
    features: &Features,
    resolver: &mut dyn ResolverExpand,
)

Traverse the crate, collecting all the test functions, eliding any existing main functions, and synthesizing a main test harness