rustc_builtin_macros::util
Function check_zero_tts

pub(crate) fn check_zero_tts(
    cx: &ExtCtxt<'_>,
    span: Span,
    tts: TokenStream,
    name: &str,
)

Non-fatally assert that tts is empty. Note that this function returns even when tts is non-empty, macros that need to stop compilation should call cx.diagnostic().abort_if_errors() (this should be done as rarely as possible).