rustc_builtin_macros
Module deriving

The compiler code necessary to implement the #[derive] extensions.

Modules

bounds ğŸ”’

clone ğŸ”’

debug ğŸ”’

decodable ğŸ”’

The compiler code necessary for #[derive(RustcDecodable)]. See encodable.rs for more.

default ğŸ”’

encodable ğŸ”’

The compiler code necessary to implement the #[derive(RustcEncodable)] (and RustcDecodable, in decodable.rs) extension. The idea here is that type-defining items may be tagged with #[derive(RustcEncodable, RustcDecodable)].

eq ğŸ”’

generic ğŸ”’

Some code that abstracts away much of the boilerplate of writing derive instances for traits. Among other things it manages getting access to the fields of the 4 different sorts of structs and enum variants, as well as creating the method and impl ast instances.

hash ğŸ”’

ord ğŸ”’

partial_eq ğŸ”’

partial_ord ğŸ”’

smart_ptr ğŸ”’

Macros

path_local ğŸ”’

path_std ğŸ”’

pathvec_std ğŸ”’

Structs

BuiltinDerive ğŸ”’

Functions

assert_ty_bounds ğŸ”’

call_intrinsic ğŸ”’

Constructs an expression that calls an intrinsic

call_unreachable ğŸ”’

Constructs an expression that calls the unreachable intrinsic.

Type Aliases

BuiltinDeriveFn ğŸ”’