Module rustc_builtin_macros::format

Modules

ast 🔒

expand 🔒

Enums

PositionUsedAs 🔒

Functions

expand_format_args

expand_format_args_impl 🔒

expand_format_args_nl

invalid_placeholder_type_error 🔒

make_format_args

parse_args 🔒

Parses the arguments from the given list of tokens, returning the diagnostic if there’s a parse error so we can continue parsing other format! expressions.

report_invalid_references 🔒

Handle invalid references to positional arguments. Output different errors for the case where all arguments are positional and for when there are named arguments or numbered positional arguments in the format string.

report_missing_placeholders 🔒