rustc_builtin_macros
Module source_util

Functions

expand_column 🔒

expand_file 🔒

file!(): expands to the current filename */ The source_file (loc.file) contains a bunch more information we could spit out if we wanted.

expand_include 🔒

include! : parse the given file as an expr This is generally a bad idea because it’s going to behave unhygienically.

expand_include_bytes 🔒

expand_include_str 🔒

include_str!: read the given file, insert it as a literal string expr

expand_line 🔒

line!(): expands to the current line number

expand_mod 🔒

expand_stringify 🔒

find_path_suggestion 🔒

load_binary_file 🔒