Module rustc_builtin_macros::test_harness

Structs

EntryPointCleaner ğŸ”’

A folder used to remove any entry points (like fn main) because the harness coroutine will provide its own

InnerItemLinter ğŸ”’

Test ğŸ”’

TestCtxt ğŸ”’

TestHarnessGenerator ğŸ”’

Functions

entry_point_type ğŸ”’

generate_test_harness ğŸ”’

Crawl over the crate, inserting test reexports and the test main function

get_test_name ğŸ”’

get_test_runner ğŸ”’

inject

Traverse the crate, collecting all the test functions, eliding any existing main functions, and synthesizing a main test harness

mk_main ğŸ”’

Creates a function item for use as the main function of a test build. This function will call the test_runner as specified by the crate attribute

mk_tests_slice ğŸ”’

Creates a slice containing every test like so: &[&test1, &test2]